requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
bitarray==3.4.0
bitpay==7.0.3
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.1.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
cytoolz==1.0.1
Django==5.2
django-allauth==65.8.0
django-cors-headers==4.7.0
django-extensions==4.1
django-redis==5.4.0
django-sslserver==0.22
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
Flask==3.1.0
frozenlist==1.6.0
git-filter-repo==2.47.0
gunicorn==23.0.0
hexbytes==1.3.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
packaging==25.0
parsimonious==0.10.0
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pyunormalize
qrcode==8.1
redis==6.0.0
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
six==1.17.0
sqlparse==0.5.3
stripe==12.0.1
toolz==1.0.0
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.13
web3==7.11.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
yarl==1.20.0
python-decouple
dj-database-url
dj-rest-auth==7.0.1
whitenoise
django-csp>=3.7
django-storages
boto3
beautifulsoup4
django-filter
selenium
django-q2
